% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candles.R
\name{get_candle_durations}
\alias{get_candle_durations}
\title{Get Candle Durations-Intervals Mapping}
\usage{
get_candle_durations(...)
}
\arguments{
\item{...}{Further arguments to \link{query_iss}.}
}
\value{
A tibble with the durations-intervals mapping.
}
\description{
REST path: \verb{/index?iss.only=durations} (see http://iss.moex.com/iss/reference/28)
}
\examples{
\dontrun{
get_candle_durations()
}
}
