% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.R
\name{get_security_info}
\alias{get_security_info}
\title{Get a Security's Info}
\usage{
get_security_info(secid, ...)
}
\arguments{
\item{secid}{A security ID.}

\item{...}{Further arguments to \link{query_iss}}
}
\value{
A list with two tibbles:
\itemize{
\item \code{description}: Full instrument profile incl. its type, listing level, etc.
\item \code{boards}: The security's boards with \code{boardid}, \code{history_from}, ... columns.
}
}
\description{
REST path: \verb{securities/[security]} (see \url{http://iss.moex.com/iss/reference/13})
}
\examples{
\dontrun{
get_security_info(secid = 'SBER')
}
}
