% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MoexCandles}
\alias{plot.MoexCandles}
\title{Plot Candles}
\usage{
\method{plot}{MoexCandles}(candles_df)
}
\arguments{
\item{candles_df}{A candles tibble as returned by \link{get_candles}}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Candles
}
\examples{
\dontrun{
get_candles(secid = 'SBER', from = '2020-01-01') |>
    plot()
}
}
