% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.R
\name{search_security}
\alias{search_security}
\title{Search for a Single Security}
\usage{
search_security(query, ...)
}
\arguments{
\item{query}{(A part of the) ID, name, ISIN, issuer ID, reg. number of a security.}

\item{...}{Further arguments to \link{query_iss}}
}
\value{
A tibble with a list of matched securities.
}
\description{
REST path: `/securities`` (See \url{http://iss.moex.com/iss/reference/5})
}
\examples{
\dontrun{
search_security(query = 'SBER')
}
}
