% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MoexCandles}
\alias{plot.MoexCandles}
\title{Plot Candles}
\usage{
\method{plot}{MoexCandles}(x, ...)
}
\arguments{
\item{x}{A candles tibble as returned by \code{\link[=get_candles]{get_candles()}}}

\item{...}{For compatibility with the generic; ignored}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Candles
}
\examples{
\dontrun{
get_candles(secid = 'SBER', from = '2020-01-01') |>
    plot()
}
}
